fs: Return ErrNotExist in ArchiveFS
Other minor fixes
