+++ archiver_test.go
@@ -113,30 +113,3 @@ func TestSkipList(t *testing.T) {
 		}
 	}
 }
-
-func TestPathWithoutTopDir(t *testing.T) {
-	for i, tc := range []struct {
-		input, expect string
-	}{
-		{
-			input:  "a/b/c",
-			expect: "b/c",
-		},
-		{
-			input:  "b/c",
-			expect: "c",
-		},
-		{
-			input:  "c",
-			expect: "",
-		},
-		{
-			input:  "",
-			expect: "",
-		},
-	} {
-		if actual := pathWithoutTopDir(tc.input); actual != tc.expect {
-			t.Errorf("Test %d (input=%s): Expected '%s' but got '%s'", i, tc.input, tc.expect, actual)
-		}
-	}
-}
