+++ fs_test.go
@@ -1,0 +1,30 @@
+package archiver
+
+import "testing"
+
+func TestPathWithoutTopDir(t *testing.T) {
+	for i, tc := range []struct {
+		input, expect string
+	}{
+		{
+			input:  "a/b/c",
+			expect: "b/c",
+		},
+		{
+			input:  "b/c",
+			expect: "c",
+		},
+		{
+			input:  "c",
+			expect: "c",
+		},
+		{
+			input:  "",
+			expect: "",
+		},
+	} {
+		if actual := pathWithoutTopDir(tc.input); actual != tc.expect {
+			t.Errorf("Test %d (input=%s): Expected '%s' but got '%s'", i, tc.input, tc.expect, actual)
+		}
+	}
+}
